refactor: io 계층 클래스를 io디렉토리로 이동
