+++ src/main/java/lotto/Application.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.io.ConsoleUserInterface;
+
 public class Application {
     public static void main(String[] args) {
         LottoMission mission = new LottoMission();
