+++ src/main/java/lotto/io/ConsoleUserInterface.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.io;
 
 import lotto.view.Prompt;
