+++ src/main/java/lotto/io/UserInputReader.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.io;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.Arrays;
